Make sure references follows API documentation
